refactor: split into diff files and restructuring
